+++ src/tbox/platform/file.c
@@ -148,4 +148,9 @@ tb_bool_t tb_file_access(tb_char_t const* path, tb_size_t mode)
     tb_trace_noimpl();
     return tb_false;
 }
+tb_bool_t tb_file_touch(tb_char_t const* path, tb_time_t atime, tb_time_t mtime)
+{
+    tb_trace_noimpl();
+    return tb_false;
+}
 #endif