+++ src/tbox/platform/windows/prefix.h
@@ -40,12 +40,21 @@
  */
 
 // FILETIME => tb_time_t
-static __tb_inline__ tb_time_t tb_filetime_to_time(FILETIME ft)
+static __tb_inline__ tb_time_t tb_filetime_to_time(FILETIME const* pft)
 {
     ULARGE_INTEGER  ui = {{0}};
-    ui.LowPart      = ft.dwLowDateTime;
-    ui.HighPart     = ft.dwHighDateTime;
+    ui.LowPart      = pft->dwLowDateTime;
+    ui.HighPart     = pft->dwHighDateTime;
     return (tb_time_t)((LONGLONG)(ui.QuadPart - 116444736000000000ull) / 10000000ul);
+}
+
+// tb_time_t => FILETIME
+static __tb_inline__ tb_void_t tb_time_to_filetime(tb_time_t tm, FILETIME* pft)
+{
+    ULARGE_INTEGER ui = {{0}};
+    ui.QuadPart = (tm * 10000000ul) + 116444736000000000ull;
+    pft->dwLowDateTime = ui.LowPart;
+    pft->dwHighDateTime = ui.HighPart;
 }
 
 // get absolute path for wchar