+++ src/demo/platform/directory.c
@@ -6,7 +6,7 @@
 /* ////////////////////////////////////////////////////////////////////////////////////////////////////
  * callback
  */
-#if 1
+#if 0
 static tb_long_t tb_directory_walk_func(tb_char_t const* path, tb_file_info_t const* info, tb_cpointer_t priv)
 {
     // check
@@ -58,8 +58,8 @@ tb_int_t tb_demo_platform_directory_main(tb_int_t argc, tb_char_t** argv)
     tb_directory_walk(argv[1]? argv[1] : current, -1, tb_true, tb_directory_walk_func, tb_null);
 #elif 0
     tb_directory_remove(argv[1]);
-#elif 0
-    tb_directory_copy(argv[1], argv[2]);
+#elif 1
+    tb_directory_copy(argv[1], argv[2], TB_FILE_COPY_LINK);
 #elif 0
     tb_directory_create(argv[1]);
 #else