+++ src/tbox/platform/directory.c
@@ -66,7 +66,7 @@ tb_void_t tb_directory_walk(tb_char_t const* path, tb_long_t recursion, tb_bool_
 {
     tb_trace_noimpl();
 }
-tb_bool_t tb_directory_copy(tb_char_t const* path, tb_char_t const* dest)
+tb_bool_t tb_directory_copy(tb_char_t const* path, tb_char_t const* dest, tb_size_t flags)
 {
     tb_trace_noimpl();
     return tb_false;