+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New features
 
 * [#190](https://github.com/tboox/tbox/pull/190): Add fs watcher
+* Add `tb_file_touch` api
 
 ### Changes
 
@@ -256,6 +257,7 @@
 ### 新特性
 
 * [#190](https://github.com/tboox/tbox/pull/190): 添加文件系统状态监视器
+* 添加 `tb_file_touch` 接口
 
 ### 改进